# PERF: PeriodArray._format_native_types