--- a/pandas/tests/indexes/period/test_formats.py
+++ b/pandas/tests/indexes/period/test_formats.py
@@ -13,7 +13,7 @@ def test_to_native_types():
     index = PeriodIndex(["2017-01-01", "2017-01-02", "2017-01-03"], freq="D")
 
     # First, with no arguments.
-    expected = np.array(["2017-01-01", "2017-01-02", "2017-01-03"], dtype="=U10")
+    expected = np.array(["2017-01-01", "2017-01-02", "2017-01-03"], dtype=object)
 
     result = index._format_native_types()
     tm.assert_numpy_array_equal(result, expected)
@@ -23,7 +23,7 @@ def test_to_native_types():
     tm.assert_numpy_array_equal(result, expected)
 
     # Make sure date formatting works
-    expected = np.array(["01-2017-01", "01-2017-02", "01-2017-03"], dtype="=U10")
+    expected = np.array(["01-2017-01", "01-2017-02", "01-2017-03"], dtype=object)
 
     result = index._format_native_types(date_format="%m-%Y-%d")
     tm.assert_numpy_array_equal(result, expected)